+++ src/main/java/com/nutrition/nutritionservice/vo/DineRecommendedRateVo.java
@@ -21,7 +21,7 @@ public class DineRecommendedRateVo extends CategoryModel<Double> implements Seri
     /**
      * 日需热量水平
      */
-    private int calorie;
+    private double calorie;
 
     /**
      * 0,未知;1,平衡;2,减脂;3,增肌
